// packages/lib/src/enums/FIGFontLayoutRule.ts
export enum FIGFontLayoutRule {
    HORIZONTAL_EQUAL_CHARACTER_SMUSHING = 1,
    HORIZONTAL_UNDERSCORE_SMUSHING = 2,
    HORIZONTAL_HIERARCHY_SMUSHING = 4,
    HORIZONTAL_OPPOSITE_PAIR_SMUSHING = 8,
    HORIZONTAL_BIG_X_SMUSHING = 16,
    HORIZONTAL_HARDBLANK_SMUSHING = 32,
    HORIZONTAL_KERNING = 64,
    HORIZONTAL_SMUSHING = 128,
    VERTICAL_EQUAL_CHARACTER_SMUSHING = 256,
    VERTICAL_UNDERSCORE_SMUSHING = 512,
    VERTICAL_HIERARCHY_SMUSHING = 1024,
    VERTICAL_HORIZONTAL_LINE_SMUSHING = 2048,
    VERTICAL_VERTICAL_LINE_SUPERSMUSHING = 4096,
    VERTICAL_KERNING = 8192,
    VERTICAL_SMUSHING = 16384
}